deleted CollectSignersSignaturesUseCase and moved SignRequest to domain/model package
